Add ReactCursor.js to this project
I hesitated adding this to wingspan-cursor, since it didn't depend on
react before and I didn't think that was the place or reason to
introduce a react dependency.

Note the utter absence of any mention of a clone function. We're
depending completely on React.addons.update to do its magic for us.